requirements.txt
aiogram==3.3.0
aiohttp==3.9.5
APScheduler==3.10.4
aiosqlite==0.19.0
python-dotenv==1.0.1
